change strict user test case
